Correctly release connections on error cases
There was a potential race condition between connection acquisition,
error handling, and release: if an error occurred during acquisition it
was possible that we wouldn't release the connection back to the pool,
resulting in slow pool drain. This manifested as an inability to acquire
a client instance (nonstop AcquisitionTimeoutError) at some eventual
point post-deploy.