Only wrap actual RPCs
Do a bit of work to figure out exactly which methods on a Thrift object
are actual RPCs and wrap those in pooledThriftClient, creating a far
less confusing client object.

This is more annoying than it should be because the generated thrift
definitions don't contain a lot of information about what is an RPC and
what isn't (e.g., a simple list). Instead, since it emits for each RPC
`do_thing` helper methods `send_do_thing` and `recv_do_thing`, use the
existence of those to filter.